+++ mythtv/external/libmythsoundtouch/cpu_detect_x86_gcc.cpp
@@ -95,7 +95,7 @@ static int mm_support(void)
     int rval = 0;
     int eax, ebx, ecx, edx;
     int max_std_level, max_ext_level, std_caps=0, ext_caps=0;
-    long a, c;
+    intptr_t a, c;
 
     __asm__ __volatile__ (
                           /* See if CPUID instruction is supported ... */
